requirements.txt
opencv-python 
google-genai 
pydantic 
sounddevice 
google-cloud-speech
dotenv
elevenlabs
